Added services for the client with his template

+++ car_rental/src/main/java/com/projecttuto/vehicule_rental/enums/SubscriptionType.java
@@ -15,9 +15,9 @@ public enum SubscriptionType {
 
     private final double price;
 
-    private final int reduction;
+    private final double reduction;
 
-    SubscriptionType(double price,int reduction){
+    SubscriptionType(double price,double reduction){
         this.price=price;
         this.reduction=reduction;
     }
